Ohhhhhhhhhhh, I figured it out. I decided to try to maybe expand the macro to see what was going on, and using https://sourceware.org/gdb/current/onlinedocs/gdb/Macros.html, used GDB to expand it. This was the result:
{
    res = 1;
    int* b = a;
    for(int* b2 = b; b != e1; ++b, ++b2) {
        printf("OwO: %d | %d\n", *b, *b2);
        if(*b != *b2) {
	    res = 0;
            break;
        }
    }
}

The b being set as a temporary throw-away within the macro was
essentially not only shadowing the original, but making it so that the b
being passed in was a referral to this temporary variable. I added a few
more underscores to make the difference explicit, and it seems to be
passing now!