Victory! This took some time to think through. Pasting notes below:
[a, a + n]

a, i (where it is the ptr)

0 1 2 3
x x x 0
      i

how to insert?

increase by 1 capacity if needed
move [i, end) to [i + 1, end + i)
(end, i)
for(iter it = end - 1; it != i; --it) {
   *(it + 1) = *it;
}
*(it + 1) = *it;

Now, insert at the pos:
*i = val;

Realization: iterators are invalidated anyways. However, what does that
leave us on...? Why does this function take an iterator? Convenience? We
have to recalculate it anyways... so, could just give us the position as
size_t to begin with... working now, though we still have some kinks

===============
NEW THINKING
===============

Ahhh - this algorithm is bad if alloc = 0... nuts

0
0

Let's say, we increase by one:

0 1
0 g

v->avail = 0;
v->tail = 1;

0 1 2 3
x x 0

Imagining insertion at one, where v->avail = 2

Traversal: move everything from [pos, avail) to [pos + 1, avail+1)
However, we need to traverse backwards, so (avail, pos] to (avail + 1,
pos + 1]
Also, It's possible for avail == pos == 0.

Maybe so for(iter it = avail; it != pos; --it)
Then, we make the shifts relative to the end? How do we prove this?
(avail, pos] == [avail, pos) ???
Ahhh, well, they are equivalent in terms of size, I suppose, the
ranges...

For each, we want to shift with out (avail, pos] to (avail + 1, pos + 1]
above though

The first element in (avail, pos] is avail - 1. In [avail, pos), the
equivalent... is also avail - 1... uhhh... *(it) = *(it - 1)?

The last case is when we hit pos + 1 and do *(pos + 1) = *(pos). Is this
right?

0 1 2 3
x x 0
  pos

I suppose so...

0 1 2 3 4
x x x 0
  p p+1

0 1 2 3 4
x   x x 0
  p p+1

There should be a way to prove this kind of thing, I'm sure of it.